Separate tests of acquisition optimizers to files
